benches/rpc: rename raftstore to tikv_server
