+++ benches/rpc/mod.rs
@@ -11,4 +11,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod raftstore;
+mod tikv_server;
